+++ R/RcppExports.R
@@ -133,8 +133,8 @@ py_call_impl <- function(x, args = NULL, keywords = NULL) {
     .Call(`_reticulate_py_call_impl`, x, args, keywords)
 }
 
-py_dict <- function(keys, items, convert) {
-    .Call(`_reticulate_py_dict`, keys, items, convert)
+py_dict_impl <- function(keys, items, convert) {
+    .Call(`_reticulate_py_dict_impl`, keys, items, convert)
 }
 
 py_dict_get_item <- function(dict, key) {
@@ -191,5 +191,9 @@ py_run_file_impl <- function(file, local = FALSE, convert = TRUE) {
 
 py_eval_impl <- function(code, convert = TRUE) {
     .Call(`_reticulate_py_eval_impl`, code, convert)
+}
+
+readline <- function(prompt) {
+    .Call(`_reticulate_readline`, prompt)
 }